* code/hygiene:
  fix, do not use ternary ?: for idx assignment in prepare()
  fix, PEBKAC, paren assignments, then comparisons
  minor, mark next_aligned_stamp() unused
  minor, make sidx_t unsigned in nature
  minor, be very clear about comparisons in tz transition code
  fix, do not allow negative column numbers in prchunk_getcolno()
  minor, fix code so it reflects our signedness assumptions

+++ src/darray.c
@@ -228,7 +228,7 @@ darray_fmread(fmcmb_t stream)
 	if (fm_read_int32(stream, &d->num_cells) < 0) {
 		goto exit_da_created;
 	}
-	if (d->num_cells > SIZE_MAX / sizeof(*d->cells)) {
+	if ((size_t)d->num_cells > SIZE_MAX / sizeof(*d->cells)) {
 		goto exit_da_created;
 	}
 	if ((d->cells = malloc(d->num_cells * sizeof(*d->cells))) == NULL) {
